feat: 카테고리 컴포넌트로 분리
