+++ frontend/src/views/user/HomeView.vue
@@ -1,50 +1,8 @@
 <template>
   <div class="main-page">
     <CommonBannerVue />
-    <!-- <div class="banner">
-      <button @click="onClickBannerBtn('prev')">
-        <span class="material-icons-outlined"> arrow_back_ios </span>
-      </button>
-      <div class="banner-img">
-        <img :src="bannerList[viewBanner]" alt="배너" />
-      </div>
-      <button @click="onClickBannerBtn('next')">
-        <span class="material-icons-outlined"> arrow_forward_ios </span>
-      </button>
-    </div> -->
     <div class="main">
-      <div class="category">
-        <button @click="onClickPrevCategory">
-          <span class="material-icons-outlined"> arrow_back_ios </span>
-        </button>
-        <ul>
-          <li v-for="category in viewCategory" :key="category.id">
-            <button
-              :class="category.id === selectedCategoryId ? `selected` : ``"
-              @click="onClickCategory(category.id)"
-            >
-              {{ category.name }}
-            </button>
-          </li>
-        </ul>
-        <button @click="onClickNextCategory">
-          <span class="material-icons-outlined"> arrow_forward_ios </span>
-        </button>
-      </div>
-      <div class="category small-category">
-        <ul v-if="smallCategoryIsOpen">
-          <li v-for="category in categories" :key="category.id">
-            <button
-              :class="category.id === selectedCategoryId ? `selected` : ``"
-              @click="onClickCategory(category.id)"
-            >
-              {{ category.name }}
-            </button>
-          </li>
-        </ul>
-        <div v-else>카테고리 열기</div>
-        <button @click="openCategory">{{ smallCategoryIsOpen ? `닫기` : `열기` }}</button>
-      </div>
+      <CategoryList v-model:selectedCategory="selectedCategory" />
       <div class="search-bar">
         <input type="text" id="search" @keyup.enter="onClickSearch" />
         <label @click="onClickSearch">
@@ -64,96 +22,14 @@
 </template>
 
 <script setup lang="ts">
-import { computed, ref } from 'vue';
+import { ref, watchEffect } from 'vue';
 import image from '@/assets/tmp/images/image.png';
 import router from '@/router';
 import ProductCardVue from '@/components/ProductCard.vue';
 import CommonBannerVue from '@/components/CommonBanner.vue';
-import type { category } from '@/types/product';
+import CategoryList from '@/components/CategoryList.vue';
 
-// TODO: 수정 -> 서버로부터
-const categories = ref<Array<category>>([
-  {
-    id: 1,
-    name: '전체'
-  },
-  {
-    id: 2,
-    name: '2'
-  },
-  {
-    id: 3,
-    name: '3'
-  },
-  {
-    id: 4,
-    name: '4'
-  },
-  {
-    id: 5,
-    name: '5'
-  },
-  {
-    id: 6,
-    name: '6'
-  },
-  {
-    id: 7,
-    name: '7'
-  },
-  {
-    id: 8,
-    name: '8'
-  },
-  {
-    id: 9,
-    name: '9'
-  },
-  {
-    id: 10,
-    name: '10'
-  },
-  {
-    id: 11,
-    name: '11'
-  },
-  {
-    id: 12,
-    name: '12'
-  },
-  {
-    id: 13,
-    name: '13'
-  },
-  {
-    id: 14,
-    name: '14'
-  },
-  {
-    id: 15,
-    name: '15'
-  },
-  {
-    id: 16,
-    name: '16'
-  },
-  {
-    id: 17,
-    name: '17'
-  },
-  {
-    id: 18,
-    name: '18'
-  },
-  {
-    id: 19,
-    name: '19'
-  },
-  {
-    id: 20,
-    name: '20'
-  }
-]);
+const selectedCategory = ref('ALL');
 
 const products = ref([
   {
@@ -193,40 +69,9 @@ const products = ref([
   }
 ]);
 
-const categoryPageNumber = ref(0);
-const viewCategoryNumber = ref(8);
-const selectedCategoryId = ref(1);
-
-const smallCategoryIsOpen = ref(false);
-
-const openCategory = () => {
-  smallCategoryIsOpen.value = !smallCategoryIsOpen.value;
-};
-
-const onClickPrevCategory = () => {
-  categoryPageNumber.value -= 1;
-  if (categoryPageNumber.value < 0) categoryPageNumber.value = 0;
-};
-
-const onClickNextCategory = () => {
-  categoryPageNumber.value += 1;
-  const maxCategoryPage = Math.floor(categories.value.length / viewCategoryNumber.value);
-  if (maxCategoryPage < categoryPageNumber.value) categoryPageNumber.value = maxCategoryPage;
-};
-
 const onClickSearch = () => {
   // TODO: 상품 검색
 };
-
-const onClickCategory = (id: number) => {
-  selectedCategoryId.value = id;
-  // TODO: 카테고리별 상품 출력
-};
-
-const viewCategory = computed(() => {
-  const start = categoryPageNumber.value * viewCategoryNumber.value;
-  return categories.value.slice(start, start + viewCategoryNumber.value);
-});
 
 const onClickAddProduct = () => {
   router.push('/product/add');
@@ -235,6 +80,11 @@ const onClickAddProduct = () => {
 const onClickProductCard = (id: number) => {
   router.push(`product/${id}`);
 };
+
+watchEffect(() => {
+  console.log(selectedCategory.value);
+  // TODO: 카테고리별 상품
+});
 </script>
 
 <style scoped>
@@ -242,94 +92,12 @@ const onClickProductCard = (id: number) => {
   width: 100%;
   height: 100%;
 }
-/* .banner {
-  box-sizing: content-box;
-  max-width: 1300px;
-  height: 300px;
-  position: relative;
-  display: flex;
-  align-items: center;
-  border: 1px solid rgba(109, 109, 109, 0.155);
-}
-.banner > button {
-  position: absolute;
-  background-color: transparent;
-}
-.banner > button:first-child {
-  left: 0;
-}
-.banner > button:last-child {
-  right: 0;
-}
-.banner-img {
-  width: 100%;
-  height: 100%;
-  text-align: center;
-  background-color: #198bf3;
-}
-.banner-img > img {
-  height: 100%;
-  object-fit: cover;
-} */
 .main {
   width: 100%;
   height: 100%;
   margin-top: 24px;
 }
 
-.category {
-  width: 100%;
-  padding: 20px;
-
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  gap: 12px;
-}
-.category > button {
-  font-size: 16px;
-  padding: 12px;
-}
-.category > button:active {
-  opacity: 0.4;
-}
-.category > ul {
-  width: 950px;
-  /* background-color: var(--category-bg); */
-  /* border: 1px solid rgba(135, 135, 135, 0.32); */
-  /* border-radius: 32px; */
-
-  display: flex;
-  flex-wrap: nowrap;
-  gap: 12px;
-  justify-content: center;
-
-  padding: 12px 24px;
-}
-.category > ul > li {
-  width: fit-content;
-}
-.category > ul > li > button {
-  background-color: var(--category-item-bg);
-
-  width: 100px;
-  padding: 16px 0;
-
-  border: 1px solid rgba(66, 66, 66, 0.605);
-  border-radius: 12px;
-
-  font-family: 'LINESeedKR-Bd';
-  font-size: 16px;
-}
-
-.category > ul > li > .selected {
-  background-color: var(--category-item-point-bg);
-  color: var(--category-item-point-text);
-  /* box-shadow: 1px 1px 10px rgba(176, 176, 176, 0.827); */
-}
-.small-category {
-  display: none;
-}
 .search-bar {
   position: relative;
   width: 100%;
